// src/generated/prisma/enums.ts
export const UserRole = {
  USER: 'USER',
  SUPER_ADMIN: 'SUPER_ADMIN',
  CONTENT_MANAGER: 'CONTENT_MANAGER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const AccountStatus = {
  ACTIVE: 'ACTIVE',
  DEACTIVATED: 'DEACTIVATED',
  PENDING_DELETION: 'PENDING_DELETION'
} as const

export type AccountStatus = (typeof AccountStatus)[keyof typeof AccountStatus]